% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable.R
\name{avg_light}
\alias{avg_light}
\title{Summary Table of average light sensor data}
\usage{
avg_light(
  start = min(as.Date(substring(moon_light$YYYY.MM.DDTHH.mm.ss.fff, 1, 10))),
  end = max(as.Date(substring(moon_light$YYYY.MM.DDTHH.mm.ss.fff, 1, 10))),
  ...
)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
returns a summary table and displays it
}
\description{
Summary Table of average light sensor data
}
\details{
This function takes in a time interval and will return a summary table for light data each day in the interval
}
\examples{
avg_light("2025-03-01", "2025-03-10")
}
\author{
Molly Daniel
}
