% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable.R
\name{light_by_moon}
\alias{light_by_moon}
\title{Summary Table of moon metric and light sensor data}
\usage{
light_by_moon(metric = Category)
}
\arguments{
\item{metric}{Category, Area, Phase - passed in without quotes}
}
\value{
returns a summary table and displays it
}
\description{
Summary Table of moon metric and light sensor data
}
\details{
This function returns general data about how light varies depending on moon metrics
}
\examples{
light_by_moon_phase()
}
\author{
Molly Daniel
}
