% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonLightVisualizer.R
\name{moon_light_visual}
\alias{moon_light_visual}
\title{Visual Analysis of moon phase and light sensor data}
\usage{
moon_light_visual(start_date, ...)
}
\arguments{
\item{start_date}{The desired start date of visual analysis}

\item{...}{Additional arguments}
}
\value{
returns a graph and displays it
}
\description{
Visual Analysis of moon phase and light sensor data
}
\details{
This function takes in a desired start date and will return a visualization for a week's worth of data
}
\examples{
moon_light_visual("")
}
\author{
Cameron Darling
}
